Move helper functions into separate module
